requirements.txt
flask==3.0.2
requests==2.31.0
pycryptodome==3.20.0
protobuf==4.25.3